start new `re_dev_tools` tool/crate to unify all dev tools and move re_build_examples in
