+++ Cargo.toml
@@ -29,15 +29,15 @@ version = "0.16.0-alpha.1+dev"
 # In particular: if we compile rerun 0.3.0-alpha.0 we only want it to use
 # re_log_types 0.3.0-alpha.0, NOT 0.3.0-alpha.4 even though it is newer and semver-compatible.
 re_analytics = { path = "crates/re_analytics", version = "=0.16.0-alpha.1", default-features = false }
-re_build_search_index = { path = "crates/re_build_search_index", version = "=0.16.0-alpha.1", default-features = false }
-re_build_examples = { path = "crates/re_build_examples", version = "=0.16.0-alpha.1", default-features = false }
 re_build_info = { path = "crates/re_build_info", version = "=0.16.0-alpha.1", default-features = false }
+re_build_search_index = { path = "crates/re_build_search_index", version = "=0.16.0-alpha.1", default-features = false }
 re_build_tools = { path = "crates/re_build_tools", version = "=0.16.0-alpha.1", default-features = false }
 re_build_web_viewer = { path = "crates/re_build_web_viewer", version = "=0.16.0-alpha.1", default-features = false }
 re_crash_handler = { path = "crates/re_crash_handler", version = "=0.16.0-alpha.1", default-features = false }
 re_data_source = { path = "crates/re_data_source", version = "=0.16.0-alpha.1", default-features = false }
 re_data_store = { path = "crates/re_data_store", version = "=0.16.0-alpha.1", default-features = false }
 re_data_ui = { path = "crates/re_data_ui", version = "=0.16.0-alpha.1", default-features = false }
+re_dev_tools = { path = "crates/re_dev_tools", version = "=0.16.0-alpha.1", default-features = false }
 re_entity_db = { path = "crates/re_entity_db", version = "=0.16.0-alpha.1", default-features = false }
 re_error = { path = "crates/re_error", version = "=0.16.0-alpha.1", default-features = false }
 re_format = { path = "crates/re_format", version = "=0.16.0-alpha.1", default-features = false }
@@ -47,9 +47,9 @@ re_log_encoding = { path = "crates/re_log_encoding", version = "=0.16.0-alpha.1"
 re_log_types = { path = "crates/re_log_types", version = "=0.16.0-alpha.1", default-features = false }
 re_memory = { path = "crates/re_memory", version = "=0.16.0-alpha.1", default-features = false }
 re_query = { path = "crates/re_query", version = "=0.16.0-alpha.1", default-features = false }
-re_query2 = { path = "crates/re_query2", version = "=0.16.0-alpha.1", default-features = false }
 re_query_cache = { path = "crates/re_query_cache", version = "=0.16.0-alpha.1", default-features = false }
 re_query_cache2 = { path = "crates/re_query_cache2", version = "=0.16.0-alpha.1", default-features = false }
+re_query2 = { path = "crates/re_query2", version = "=0.16.0-alpha.1", default-features = false }
 re_renderer = { path = "crates/re_renderer", version = "=0.16.0-alpha.1", default-features = false }
 re_sdk = { path = "crates/re_sdk", version = "=0.16.0-alpha.1", default-features = false }
 re_sdk_comms = { path = "crates/re_sdk_comms", version = "=0.16.0-alpha.1", default-features = false }
