+++ crates/re_dev_tools/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "re_build_examples"
+name = "re_dev_tools"
 authors.workspace = true
-description = "Build rerun example RRD files"
+description = "Various tools for rerun development"
 edition.workspace = true
 homepage.workspace = true
 include.workspace = true
@@ -23,8 +23,18 @@ re_format.workspace = true
 # External
 anyhow.workspace = true
 argh.workspace = true
+camino.workspace = true
+cargo_metadata.workspace = true
+glob.workspace = true
 indicatif.workspace = true
+itertools.workspace = true
+roxmltree = "0.19.0"
 rayon.workspace = true
+rustdoc-json = "0.8.8"
+rustdoc-types = "0.24.0"
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
+serde_yaml.workspace = true
 toml = { workspace = true, features = ["parse", "preserve_order"] }
+ureq = { workspace = true, features = ["json"] }
+url.workspace = true
