+++ docs/snippets/snippets.toml
@@ -1,4 +1,4 @@
-# This file is read in `compare_snippet_output.py` and the `snippets` command in `re_build_examples`.
+# This file is read in `compare_snippet_output.py` and the `snippets` command in `re_dev_tools`/`build_examples`.
 
 # These entries won't run at all.
 #
